backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.7
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
gevent==24.2.1
greenlet==3.0.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
python-dateutil==2.9.0.post0
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
rpds-py==0.19.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.4.post2
